add: channel labels
