+++ src/view/show_mode/node_editor_widgets/cue_editor/view_settings.py
@@ -1,0 +1,1 @@
+CHANNEL_DISPLAY_HEIGHT = 64
